#59 Changes up to API v1.24 (#60)
* Changes up to API v1.24
* Changed all floats (single) to double
* Made models related to splashsettings nullable.
* Updated changelog and version file

+++ Meraki.Api/Data/SwitchPortStatus.cs
@@ -40,7 +40,7 @@ public class SwitchPortStatus
 	public int ClientCount { get; set; }
 
 	[DataMember(Name = "powerUsageInWh")]
-	public float PowerUsageInWh { get; set; }
+	public double PowerUsageInWh { get; set; }
 
 	[DataMember(Name = "trafficInKbps")]
 	public TrafficInKbps TrafficInKbps { get; set; } = new();
